creating a new container for new questionnaire form, allowing normal browser navigation between pages
